Changing photomise to use a more structured cli approach
