+++ photomise.egg-info/PKG-INFO
@@ -1,0 +1,4 @@
+Metadata-Version: 2.2
+Name: photomise
+Version: 0.1.0
+License-File: LICENSE
