+++ photomise.egg-info/SOURCES.txt
@@ -1,0 +1,17 @@
+LICENSE
+README.md
+setup.py
+photomise/__init__.py
+photomise.egg-info/PKG-INFO
+photomise.egg-info/SOURCES.txt
+photomise.egg-info/dependency_links.txt
+photomise.egg-info/entry_points.txt
+photomise.egg-info/top_level.txt
+photomise/cli/__init__.py
+photomise/cli/main.py
+photomise/cli/post/__init__.py
+photomise/cli/post/atprotocol.py
+photomise/cli/settings/__init__.py
+photomise/cli/settings/filters.py
+photomise/cli/settings/interactive.py
+photomise/cli/settings/locations.py
