+++ photomise/__init__.py
@@ -1,0 +1,7 @@
+"""Photomise - Photo processing and social media posting tool"""
+
+from typing import Final
+
+__version__: Final[str] = "0.1.0"
+__author__: Final[str] = "Jesse Fuller"
+__license__: Final[str] = "MIT"
