+++ photomise/utilities/atprotocol.py
@@ -1,0 +1,2 @@
+def fix_bluesky_username(current):
+    return current.replace("@", "").strip()
