+++ photomise/utilities/constants.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+CONFIG_FILE = "config.ini"
+SHARED_DB_PATH = "shared.json"
+BLUESKY_SERVICE_NAME = "photomise-atprotocol-bluesky"
+APP_DIR = Path.home() / ".photomise"
+LOG_DIR = APP_DIR / "logs"
